requirements.txt
flask==2.3.3
google-generativeai==0.3.1
PyPDF2==3.0.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7